file_storage/requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiomysql==0.2.0
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
bcrypt==4.1.2
black==24.4.2
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
databases==0.9.0
Django==4.2.14
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-admin==1.0.4
fastapi-cache2==0.2.1
fastapi-cli==0.0.4
fastapi-redis==0.1.11
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
fastui==0.6.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
iso8601==1.1.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
makefun==1.15.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
orjson==3.10.3
packaging==24.1
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pwdlib==0.2.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pypika-tortoise==0.1.6
pytest==8.2.2
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqladmin==0.18.0
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
starlette==0.37.2
time-machine==2.14.2
tortoise-orm==0.21.4
typer==0.12.3
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
WTForms==3.1.2
yarl==1.9.4
aiobotocore
passlib
asyncpg